backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-socketio==5.10.0
python-dotenv==1.0.1
websockets==12.0
python-engineio==4.8.0
aiohttp==3.9.1
pydantic==2.6.1 
